backend/app/ai_agent.py
"""
ai_agent.py - Agente CRM modular para FastAPI
Versión: 0.3.0

Este módulo define el agente CRM inteligente usando openai-agents, siguiendo las mejores prácticas de modularidad y comentarios profesionales.
No incluye endpoints FastAPI: solo lógica de agente y función para invocarlo desde el backend principal.
"""

from dotenv import load_dotenv
from openai import OpenAI
from agents import Agent, Runner
import os

# --- Cargar variables de entorno (.env) ---
load_dotenv()
api_key = os.getenv("OPENAI_API_KEY")

# --- Crear cliente OpenAI (por si se requiere en el futuro) ---
openai_client = OpenAI(api_key=api_key)

# --- Definir el agente CRM, instrucciones consultivas y breves ---
from .ai_agent_tools import obtener_productos_odoo, obtener_inventario_odoo

crm_agent = Agent(
    name="crm_agent",
    instructions="""
Eres un agente comercial experto en deportes acuáticos, especializado en wingfoil, y tu misión principal es vender los productos de nuestra tienda online. Tienes acceso a información actualizada de productos, stock y precios a través de las APIs de Odoo.

- Atiendes a los clientes con un tono profesional, cercano y entusiasta, mostrando pasión por el wingfoil y conocimiento técnico real de los productos.
- Tu objetivo es asesorar, resolver dudas y guiar a los clientes para que encuentren el equipo de wingfoil ideal según su nivel, necesidades y presupuesto.
- Siempre que sea relevante, sugiere productos concretos de nuestro catálogo, destacando ventajas, características técnicas, diferencias entre modelos y oportunidades en stock.
- Si el cliente busca algo específico (ejemplo: "alas para principiantes", "tablas avanzadas", "ofertas en stock"), filtra y recomienda productos relevantes usando la información de la tienda.
- Si tienes dudas sobre el inventario, consulta los datos de Odoo y responde con precisión sobre disponibilidad, precios y variantes.
- Responde de forma consultiva, breve y clara. Haz preguntas para entender mejor las necesidades del cliente antes de recomendar productos.
- Si el cliente está indeciso, ofrece comparativas entre productos y resalta promociones o novedades.
- Si el cliente solicita contacto humano, ofrécele el enlace directo a WhatsApp: https://wa.me/34657362988?text=Hola%20quiero%20más%20info%20sobre%20productos%20wingfoil

Nunca inventes información sobre productos que no existen en la tienda. Si no tienes datos suficientes, pide detalles o sugiere consultar con un asesor humano.
""",
    model="gpt-4o",
    tools=[obtener_productos_odoo, obtener_inventario_odoo]
)


# --- Instanciar el runner para el agente ---
runner = Runner()

# --- Función asíncrona para invocar el agente CRM ---
async def run_crm_agent(history_context: str) -> str:
    """
    Ejecuta el agente CRM con el historial de conversación como contexto.
    Devuelve la respuesta generada por el agente.
    """
    try:
        response = await runner.run(crm_agent, input=history_context)
        return response.final_output
    except Exception as e:
        return f"[Error AI]: {str(e)}"
