backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
openai==1.76.0
openai-agents==0.0.13
httpx==0.27.0
